pystray_sample/requirements.txt
pystray==0.19.4
Pillow==9.2.0
